chore(client): split timeouts module into a file
